Split app.js phase 3: extract schedule-engine, persistence, render hubs
First import cycles appear here (render <-> app.js); safe because all
cross-module calls are hoisted function declarations invoked at runtime,
never during module evaluation. Cross-module mutable vars
(lastBlockedDomains, activatedBlockIds, pauseScheduleData, isScheduleMode)
moved to state.js. In-app tests pass (132/132).
